requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
Pillow==12.1.1
Jinja2==3.1.6
python-multipart==0.0.22
